src/main/java/pl/great/waw/calculations/Product.java
package pl.great.waw.calculations;

import java.math.BigDecimal;

public class Product {

    private String name = "";
    private final BigDecimal price;

    public Product(String name, BigDecimal price) {
        this.name = name;
        this.price = price;
    }

    public String getName() {
        return name;
    }

    public BigDecimal getPrice() {
        return price;
    }

    @Override
    public String toString() {
        return "Product {" +
                "name='" + name + '\'' +
                ", price=" + price +
                '}';
    }
}
